Fork Jackcess 4.0.6-SNAPSHOT
Fork project Jackcess, latest unreleased version 4.0.6-SNAPSHOT as of 2024-02-12 from:

https://github.com/jahlborn/jackcess
(Github mirror of http://jackcess.sourceforge.net/)

The intention of the fork is to provide a modernized version of Jackcess that requires
Java 11 minimum and has as few as possible external dependencies.

The original developers or contributors of Jackcess are:

- Tim McCune <javajedi@users.sf.net>
- James Ahlborn <jtahlborn@yahoo.com>
- Dan Rollo <bhamail@users.sf.net>

The project is licensed under Apache-2.0 license at time of fork.
The fork fully honors the requirements of this license.